Omit ':{}' when pretty-printing Dependency with LMainLibName sublib.
The current output when pretty-printing a Dependency is to unconditionally
include :{<sublibs>}, even when sub-libs is the singleton LMainLibName. This to
begin with is an experimental syntax (#5660) and issues a warning if
cabal-version is <3.0. Since 'cabal format' and soon 'cabal init' use
'Dependency -> PrettyField a' this produces contents in the .cabal file which
cannot be parsed for older config versions.

Testing:

- Verified that round-trip tests through parser/pretty pass
- Unit tests for sublibs={} and sublibs={LMainLibName}